+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased (v0.3.1)
+* Fix some cases where ffmpeg progress & VMAF score output parsing failed.
 * Fix some edge cases where crf-search would succeed exceeding the specified `--max-encoded-percent`.
 
 # v0.3.0